200x
Turns out python really hates editing long lists. Preprocessing now runs like a breeze. The trick was to make a new list and work on sub lists that all stitch together to a final list.
This could probably be optimized further... if i should ever need to work with corpi in the high millions.